+++ CsharpRPG/Engine/World/Dungeon.cs
@@ -6,19 +6,22 @@ namespace RPG.Engine
     public class Dungeon // Class holding the play area and tiles inside
     {
         public string Name { get; set; }
+        public List<Location> Rooms { get; set; }
+
         Random rand = new Random();
 
-        public Dungeon(string _name, int numOfRooms, Biome biome)
+        public Dungeon(string _name, int numOfRooms, Biome biome, int playerX, int playerY)
         {
             Name = _name;
-            BuildMap(_name, numOfRooms, biome);
+            BuildMap(numOfRooms, biome, playerX, playerY);
         }
 
-        void BuildMap(string name, int numOfRooms, Biome biome)
+        void BuildMap(int numOfRooms, Biome biome, int playerX, int playerY)
         {
+            Rooms = new List<Location>();
             for (int i = 0; i < numOfRooms; i++)
             {
-                GenerateRoom(biome, i, "", "", rand.Next(5, 10), rand.Next(5, 10), 0, 0, "South");
+                GenerateRoom(biome, i, "", "", rand.Next(5, 10), rand.Next(5, 10),ref playerX, ref playerY, "South");
             }
         }
 
@@ -30,7 +33,8 @@ namespace RPG.Engine
         // --- Chest Tile [4]
         // --- South Facing Door Tile [5]
         // --- North Facing Door Tile [6]
-        // --- EastWest Facing Door Tile [7]
+        // --- East Facing Door Tile [7]
+        // --- West Facing Door Tile [8]
 
         /// <summary>
         /// Create a Procedurally generated location
@@ -45,41 +49,62 @@ namespace RPG.Engine
         /// <param name="playerY">the last y coordinate of the player</param>
         /// <param name="playerFacing">the last direction the player was facing</param>
         /// <returns>Void</returns>
-        public void GenerateRoom (Biome biome, int id, string name, string desc, int width, int length, int playerX, int playerY, string playerFacing)
+        public void GenerateRoom (Biome biome, int id, string name, string desc, int width, int length, ref int playerX, ref int playerY, string playerFacing)
         {
             bool containsLoot = false, hasMonster = false, hasDoor = false;
 
             // Generate a room based on the biome and tile size
-            Location procLoc = new Location(id, name, desc, width, length);
+            Rooms.Add(new Location(id, name, desc, width, length));
 
-            for(int y = 0; y < length; y++)
+            if(biome.AvailibleTiles.Count != 0)
             {
-                for (int x = 0; x < width; x++)
+                GenerateTiles(width, length, biome, Rooms.Count - 1);
+
+                // Room needs atleast one door
+                GenerateDoors(ref hasDoor, ref playerX, ref playerY,playerFacing, width, length, biome, Rooms.Count - 1);
+
+                // Room should include some harvestable item (Treasure, Plant)
+                //GenerateLoot(ref containsLoot, width, length, biome, Rooms.Count - 1);
+            }
+
+            if (biome.AvailibleMonsters.Count != 0)
+            {
+                SpawnMonster(ref hasMonster, length, width, biome, Rooms.Count - 1);
+            }
+        }
+
+        public void GenerateHallway(Biome biome, int width, int length, ref int playerX, ref int playerY, string Facing, int roomNumber)
+        {
+            for(int i = 0; i < length; i++)
+            {
+                for (int j = 0; j < width; j++)
                 {
-                    if(y == 0)
+                    if(Facing == "North" || Facing == "South")
                     {
-                        // make a South Facing wall tile
-                        procLoc.Tiles.Add(biome.AvailibleTiles[0]);
+                        if(j == 0 || j == width)
+                        {
+                            Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[2].ID, biome.AvailibleTiles[2].Name, biome.AvailibleTiles[2].Dense, i, j, biome.AvailibleTiles[2].Type));
+                        }
                     }
-                    else if(y == length - 1)
+                    if (Facing == "East" || Facing == "West")
                     {
-                        // make a North Facing wall tile
-                        procLoc.Tiles.Add(biome.AvailibleTiles[1]);
-                    }
-                    else if(x == 0 || x == width - 1)
-                    {
-                        // make a  WestEast Facing wall tile
-                        procLoc.Tiles.Add(biome.AvailibleTiles[2]);
-                    }
-                    else
-                    {
-                        // make a floor tile
-                        procLoc.Tiles.Add(biome.AvailibleTiles[3]);
+                        if (i == 0)
+                        {
+                            // make a South Facing wall tile
+                            Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[0].ID, biome.AvailibleTiles[0].Name, biome.AvailibleTiles[0].Dense, i, j, biome.AvailibleTiles[0].Type));
+                        }
+                        else if (i == length - 1)
+                        {
+                            // make a North Facing wall tile
+                            Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[1].ID, biome.AvailibleTiles[1].Name, biome.AvailibleTiles[1].Dense, i, j, biome.AvailibleTiles[1].Type));
+                        }
                     }
                 }
             }
+        }
 
-            // Room needs atleast one door
+        void GenerateDoors(ref bool hasDoor, ref int playerX, ref int playerY, string playerFacing, int width, int length, Biome biome, int roomNumber)
+        {
             while (!hasDoor)
             {
                 Random rand = new Random();
@@ -87,70 +112,122 @@ namespace RPG.Engine
                 // place door where player entered from
                 int x = playerX, y = playerY;
 
-                for (int i = 0; i < procLoc.Tiles.Count; i++)
+                for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
                 {
-                    if (procLoc.Tiles[i].X == x && procLoc.Tiles[i].Y == y)
+                    if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
                     {
-                        if(playerFacing == "South")
+                        if (playerFacing == "South")
                         {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[5], "South", null));
+                            Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[5].ID, biome.AvailibleTiles[5].Name, biome.AvailibleTiles[5].Dense, x, 0, biome.AvailibleTiles[5].Type);
+                            Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[5].ID, biome.AvailibleTiles[5].Name, biome.AvailibleTiles[5].Dense, x, 0, biome.AvailibleTiles[5].Type), "South", null));
                         }
                         else if (playerFacing == "North")
                         {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[6], "North", null));
+                            Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[6].ID, biome.AvailibleTiles[6].Name, biome.AvailibleTiles[6].Dense, x, length - 1, biome.AvailibleTiles[6].Type);
+                            Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[6].ID, biome.AvailibleTiles[6].Name, biome.AvailibleTiles[6].Dense, x, length, biome.AvailibleTiles[6].Type), "North", null));
                         }
-                        else if (playerFacing == "East" || playerFacing == "West")
+                        else if (playerFacing == "East")
                         {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[7], playerFacing, null));
+                            Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[7].ID, biome.AvailibleTiles[7].Name, biome.AvailibleTiles[7].Dense, 0, y, biome.AvailibleTiles[7].Type);
+                            Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[7].ID, biome.AvailibleTiles[7].Name, biome.AvailibleTiles[7].Dense, 0, y, biome.AvailibleTiles[7].Type), "East", null));
+                        }
+                        else if (playerFacing == "West")
+                        {
+                            Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[8].ID, biome.AvailibleTiles[8].Name, biome.AvailibleTiles[8].Dense, width - 1, y, biome.AvailibleTiles[8].Type);
+                            Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[8].ID, biome.AvailibleTiles[8].Name, biome.AvailibleTiles[8].Dense, width, y, biome.AvailibleTiles[8].Type), "West", null));
                         }
                         hasDoor = true;
                     }
                 }
+                playerX = Rooms[roomNumber].Transitions[0].X;
+                playerY = Rooms[roomNumber].Transitions[0].Y;
 
                 // Random number of doors to generate
-                int numDoors = rand.Next(width);
+                int numDoors = rand.Next(1, 3);
 
-                x = rand.Next(width);
-                y = rand.Next(length);
-
-                // EastWest facing door
-                if ((x == 0 || x == width))
+                for(int j = 0; j < numDoors; j++)
                 {
-                    for (int i = 0; i < procLoc.Tiles.Count; i++)
+                    do
                     {
-                        if (procLoc.Tiles[i].X == x && procLoc.Tiles[i].Y == y)
-                        {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[7], "East", null));
-                        }
-                    }
-                }
+                        x = rand.Next(width);
+                        y = rand.Next(length);
 
-                // South facing door
-                else if (y == 0)
-                {
-                    for (int i = 0; i < procLoc.Tiles.Count; i++)
-                    {
-                        if (procLoc.Tiles[i].X == x && procLoc.Tiles[i].Y == y)
+                        if (x == 0 || x == width || y == 0 || y == length)
                         {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[5], "South", null));
-                        }
-                    }
-                }
+                            // East facing door
+                            if (x == 0)
+                            {
+                                for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
+                                {
+                                    if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
+                                    {
+                                        if (y == 0 || y == length)
+                                        {
+                                            y = rand.Next(length);
+                                        }
+                                        Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[7].ID, biome.AvailibleTiles[7].Name, biome.AvailibleTiles[7].Dense, x, y, biome.AvailibleTiles[7].Type);
+                                        Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[7].ID, biome.AvailibleTiles[7].Name, biome.AvailibleTiles[7].Dense, x, y, biome.AvailibleTiles[7].Type), "East", null));
+                                    }
+                                }
+                            }
 
-                // North Facing Door
-                else if (y == length)
-                {
-                    for (int i = 0; i < procLoc.Tiles.Count; i++)
-                    {
-                        if (procLoc.Tiles[i].X == x && procLoc.Tiles[i].Y == y)
-                        {
-                            procLoc.Transitions.Add(new Transition(biome.AvailibleTiles[6], "North", null));
+                            // West facing door
+                            else if (x == width)
+                            {
+                                for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
+                                {
+                                    if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
+                                    {
+                                        if (y == 0 || y == length)
+                                        {
+                                            y = rand.Next(length);
+                                        }
+                                        Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[8].ID, biome.AvailibleTiles[8].Name, biome.AvailibleTiles[8].Dense, x, y, biome.AvailibleTiles[8].Type);
+                                        Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[8].ID, biome.AvailibleTiles[8].Name, biome.AvailibleTiles[8].Dense, x, y, biome.AvailibleTiles[8].Type), "West", null));
+                                    }
+                                }
+                            }
+
+                            // South facing door
+                            else if (y == 0)
+                            {
+                                for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
+                                {
+                                    if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
+                                    {
+                                        if (x == 0 || x == width)
+                                        {
+                                            x = rand.Next(width);
+                                        }
+                                        Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[5].ID, biome.AvailibleTiles[5].Name, biome.AvailibleTiles[5].Dense, x, y, biome.AvailibleTiles[5].Type);
+                                        Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[5].ID, biome.AvailibleTiles[5].Name, biome.AvailibleTiles[5].Dense, x, y, biome.AvailibleTiles[5].Type), "South", null));
+                                    }
+                                }
+                            }
+
+                            // North Facing Door
+                            else if (y == length)
+                            {
+                                for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
+                                {
+                                    if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
+                                    {
+                                        if (x == 0 || x == width)
+                                        {
+                                            x = rand.Next(width);
+                                        }
+                                        Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[6].ID, biome.AvailibleTiles[6].Name, biome.AvailibleTiles[6].Dense, x, y, biome.AvailibleTiles[6].Type);
+                                        Rooms[roomNumber].Transitions.Add(new Transition(new Tile(biome.AvailibleTiles[6].ID, biome.AvailibleTiles[6].Name, biome.AvailibleTiles[6].Dense, x, y, biome.AvailibleTiles[6].Type), "North", null));
+                                    }
+                                }
+                            }
                         }
-                    }
+                    } while (!(x == 0 || x == width || y == 0 || y == length));
                 }
             }
-
-            // Room should include some harvestable item (Treasure, Plant)
+        }
+        void GenerateLoot(ref bool containsLoot, int width, int length, Biome biome, int roomNumber)
+        {
             while (!containsLoot)
             {
                 Random rand = new Random();
@@ -158,22 +235,53 @@ namespace RPG.Engine
                 int x = rand.Next(width);
                 int y = rand.Next(length);
 
-                if(!(x == 0 || x == width))
+                if (!(x == 0 || x == width))
                 {
-                    if(!(y == 0 || y == length))
+                    if (!(y == 0 || y == length))
                     {
-                        for (int i = 0; i < procLoc.Tiles.Count; i++)
+                        for (int i = 0; i < Rooms[roomNumber].Tiles.Count; i++)
                         {
-                            if(procLoc.Tiles[i].X == x && procLoc.Tiles[i].Y == y)
+                            if (Rooms[roomNumber].Tiles[i].X == x && Rooms[roomNumber].Tiles[i].Y == y)
                             {
-                                procLoc.Tiles[i] = new Tile(biome.AvailibleTiles[4]);
+                                Rooms[roomNumber].Tiles[i] = new Tile(biome.AvailibleTiles[4]);
                                 containsLoot = true;
                             }
                         }
                     }
                 }
             }
-
+        }
+        void GenerateTiles(int width, int length, Biome biome, int roomNumber)
+        {
+            for (int y = 0; y < length; y++)
+            {
+                for (int x = 0; x < width; x++)
+                {
+                    if (y == 0)
+                    {
+                        // make a South Facing wall tile
+                        Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[0].ID, biome.AvailibleTiles[0].Name, biome.AvailibleTiles[0].Dense, x, y, biome.AvailibleTiles[0].Type));
+                    }
+                    else if (y == length - 1)
+                    {
+                        // make a North Facing wall tile
+                        Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[1].ID, biome.AvailibleTiles[1].Name, biome.AvailibleTiles[1].Dense, x, y, biome.AvailibleTiles[1].Type));
+                    }
+                    else if (x == 0 || x == width - 1)
+                    {
+                        // make a  WestEast Facing wall tile
+                        Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[2].ID, biome.AvailibleTiles[2].Name, biome.AvailibleTiles[2].Dense, x, y, biome.AvailibleTiles[2].Type));
+                    }
+                    else
+                    {
+                        // make a floor tile
+                        Rooms[roomNumber].Tiles.Add(new Tile(biome.AvailibleTiles[3].ID, biome.AvailibleTiles[3].Name, biome.AvailibleTiles[3].Dense, x, y, biome.AvailibleTiles[3].Type));
+                    }
+                }
+            }
+        }
+        void SpawnMonster(ref bool hasMonster, int length, int width, Biome biome, int roomNumber)
+        {
             // Room can include a monster spawn
             while (!hasMonster)
             {
@@ -186,14 +294,13 @@ namespace RPG.Engine
                 {
                     if (!(y == 0 || y == length))
                     {
-                        int i = rand.Next(biome.AvailibleMonsters.Count);
-                        procLoc.MonsterLivingHere = new Monster(biome.AvailibleMonsters[i]);
+                        int i = rand.Next(biome.AvailibleMonsters.Count - 1);
+                        Rooms[roomNumber].MonsterLivingHere = new Monster(biome.AvailibleMonsters[i]);
                         hasMonster = true;
                     }
                 }
             }
         }
     }
-    
 }